backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Database dependencies
sqlalchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector database
qdrant-client==1.12.1

# LLM and AI dependencies
langchain==0.3.7
langchain-community==0.3.7
openai==1.54.5
httpx==0.25.2

# Image processing for persona cards
Pillow==10.1.0

# WebSocket support
websockets==12.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pyyaml==6.0.1

# Development tools
ruff==0.1.7
mypy==1.7.1